Don't use fork to daemonize, instead use subprocess.Popen
In a highly dynamic environment like Python, forking without
immediately execcing is error-prone. There could be other
threads running, or a another currently loaded library may
require fork notifications to update internal state.

In particular some macOS library functions do not support
being called in a child after a fork. This was causing
crashing on macOS.

Unless you are in an exclusively POSIX compatible environment
that isn't running threads forking without execcing should
not be assumed to be supported.

This also is a necessary precondition to running on Windows.

This changes the API to simple_main() so will we have to bump
the major version.