Use subprocess.Popen to run FUSE main loop in background
Similar to 9a514f80b6c26151b8fab53efbdca0ca6e01674c, we cannot fork
without execcing afterwards because a Python environment generally
doesn't guarantee that is safe.

This time the fuse library was implicitly executing fork() underneath
us. The bug manifested itself when other Python threads were running
while the library called fork().  Since fork() was being called
without the GIL, if another thread was concurrently running with the
GIL, the GIL would never be released in the new process, causing a hang
when trying to reacquire the GIL in the new process.

To fix, we fold the FUSE codepath into our existing infrastructure for
running the mount in the background.